[#417] Upload each bottle built in BK separately
Problem: currently for our Big Sur CI based on Buildkite, we have
steps that check if a bottle has been built already, but the GitHub
release is never used as an intermediate repository as we still only
upload all the built bottles in a batch at the end of a run. This
works as intended for subsequent releases, but clutters the pipeline
uselessly in case of an upstream release.

Solution: changed the workflow to immediately upload each bottle on
a successful build.