[#643] Make `tezos-baking` separate package
Problem: Currently the source code of our setup and
voting wizards is in the docker/package/ directory
and it's part of the project there. This is quite confusing
and incorrect, as their code is packaged, but isn't part of the packaging logic.

Solution: Move tezos-baking sources to their own package.