[#641] Add tezos binaries tests
Problem: We want to test that latests released binaries can be installed
and work fine on different oses. Currently we support only Fedora and Ubuntu

Solution: Add github actions that run docker dontainer with given os and
install and execute each binary in it. Currently we run checks in parallel,
3 jobs per os.