[#633] Allow users to specify snapshot provider
Problem: The setup wizard currently has an option to use
the metadata provided by xtz-shots (see #441) to download
a snapshot for its node. There are however other snapshot
providers that are adopting the same metadata format, so we want
to let the users pick a different provider too.

Solution: Add the step to choose provider.