[#736] Add setup to automatically update binaries
Problem: We want to automatically update list of binaries after a tag build

Soltuion: Add setup to update list of binaries and create a followup pr in case of new release. Note that due to how it works, the first PR will probably contain a lot of changes to the list (because the list is sorted in a different order from what buildkite artifacts provide).